llvm: split codegen into modules
