+++ src/backend/llvm/compiler/codegen/instructions.rs
@@ -1,332 +1,33 @@
-use super::helpers::{as_float, as_int, map_type};
-use super::LLVMCompiler;
+use super::types::{
+    as_float, as_int, classify_tag, cmp_eq_ne_any, i64_to_ptr, map_mirtype_to_basic, map_type,
+    to_bool, to_i64_any,
+};
 use crate::backend::llvm::context::CodegenContext;
 use crate::box_trait::{BoolBox, IntegerBox, StringBox};
 use crate::boxes::{function_box::FunctionBox, math_box::FloatBox, null_box::NullBox};
-use crate::mir::function::MirModule;
+use crate::mir::function::MirFunction;
 use crate::mir::instruction::{BinaryOp, CompareOp, ConstValue, MirInstruction, UnaryOp};
-use crate::mir::ValueId;
-use inkwell::context::Context;
-use inkwell::types::BasicType; // for as_basic_type_enum()
-use inkwell::{
-    types::{BasicTypeEnum, FloatType, IntType, PointerType},
-    values::{BasicValueEnum, FloatValue, FunctionValue, IntValue, PhiValue, PointerValue},
-    AddressSpace,
-};
+use crate::mir::{BasicBlockId, ValueId};
+use inkwell::basic_block::BasicBlock;
+use inkwell::builder::Builder;
+use inkwell::types::{BasicType, BasicTypeEnum, FloatType, IntType, PointerType};
+use inkwell::values::{BasicValueEnum, FloatValue, FunctionValue, IntValue, PhiValue, PointerValue};
+use inkwell::AddressSpace;
 use std::collections::HashMap;
 
-impl LLVMCompiler {
-    pub fn new() -> Result<Self, String> {
-        Ok(Self {
-            values: HashMap::new(),
-        })
-    }
-
-    pub fn compile_module(&self, mir_module: &MirModule, output_path: &str) -> Result<(), String> {
-        if std::env::var("NYASH_CLI_VERBOSE").ok().as_deref() == Some("1") {
-            eprintln!(
-                "[LLVM] compile_module start: functions={}, out={}",
-                mir_module.functions.len(),
-                output_path
-            );
-        }
-        let context = Context::create();
-        let codegen = CodegenContext::new(&context, "nyash_module")?;
-        // Lower only Main.main for now
-        // Find entry function
-        let func = if let Some((_n, f)) = mir_module
-            .functions
-            .iter()
-            .find(|(_n, f)| f.metadata.is_entry_point)
-        {
-            f
-        } else if let Some(f) = mir_module.functions.get("Main.main") {
-            f
-        } else if let Some(f) = mir_module.functions.get("main") {
-            f
-        } else if let Some((_n, f)) = mir_module.functions.iter().next() {
-            f
-        } else {
-            return Err("Main.main function not found in module".to_string());
-        };
-
-        // Map MIR types to LLVM types via helpers
-
-        // Load box type-id mapping from nyash_box.toml (central plugin registry)
-        let box_type_ids = crate::backend::llvm::box_types::load_box_type_ids();
-
-        // Function type
-        let ret_type = match func.signature.return_type {
-            crate::mir::MirType::Void => None,
-            ref t => Some(map_type(codegen.context, t)?),
-        };
-        let fn_type = match ret_type {
-            Some(BasicTypeEnum::IntType(t)) => t.fn_type(&[], false),
-            Some(BasicTypeEnum::FloatType(t)) => t.fn_type(&[], false),
-            Some(BasicTypeEnum::PointerType(t)) => t.fn_type(&[], false),
-            Some(_) => return Err("Unsupported return basic type".to_string()),
-            None => codegen.context.void_type().fn_type(&[], false),
-        };
-        let llvm_func = codegen.module.add_function("ny_main", fn_type, None);
-
-        // Create LLVM basic blocks: ensure entry is created first to be function entry
-        let mut bb_map: HashMap<crate::mir::BasicBlockId, inkwell::basic_block::BasicBlock> =
-            HashMap::new();
-        let entry_first = func.entry_block;
-        let entry_bb = codegen
-            .context
-            .append_basic_block(llvm_func, &format!("bb{}", entry_first.as_u32()));
-        bb_map.insert(entry_first, entry_bb);
-        for bid in func.block_ids() {
-            if bid == entry_first {
-                continue;
-            }
-            let name = format!("bb{}", bid.as_u32());
-            let bb = codegen.context.append_basic_block(llvm_func, &name);
-            bb_map.insert(bid, bb);
-        }
-
-        // Position at entry
-        codegen.builder.position_at_end(entry_bb);
-
-        // SSA value map
-        let mut vmap: HashMap<ValueId, BasicValueEnum> = HashMap::new();
-
-        // Helper ops (centralized conversions and comparisons)
-        fn to_i64_any<'ctx>(
-            ctx: &'ctx Context,
-            builder: &inkwell::builder::Builder<'ctx>,
-            v: BasicValueEnum<'ctx>,
-        ) -> Result<IntValue<'ctx>, String> {
-            let i64t = ctx.i64_type();
-            Ok(match v {
-                BasicValueEnum::IntValue(iv) => {
-                    if iv.get_type().get_bit_width() == 64 {
-                        iv
-                    } else if iv.get_type().get_bit_width() < 64 {
-                        builder
-                            .build_int_z_extend(iv, i64t, "zext_i64")
-                            .map_err(|e| e.to_string())?
-                    } else {
-                        builder
-                            .build_int_truncate(iv, i64t, "trunc_i64")
-                            .map_err(|e| e.to_string())?
-                    }
-                }
-                BasicValueEnum::PointerValue(pv) => builder
-                    .build_ptr_to_int(pv, i64t, "p2i64")
-                    .map_err(|e| e.to_string())?,
-                BasicValueEnum::FloatValue(fv) => {
-                    // Bitcast f64 -> i64 via stack slot
-                    let slot = builder
-                        .get_insert_block()
-                        .and_then(|bb| bb.get_parent())
-                        .and_then(|f| f.get_first_basic_block())
-                        .map(|entry| {
-                            let eb = ctx.create_builder();
-                            eb.position_at_end(entry);
-                            eb
-                        })
-                        .unwrap_or_else(|| ctx.create_builder());
-                    let i64p = i64t.ptr_type(AddressSpace::from(0));
-                    let tmp = slot
-                        .build_alloca(i64t, "f2i_tmp")
-                        .map_err(|e| e.to_string())?;
-                    let fptr_ty = ctx.f64_type().ptr_type(AddressSpace::from(0));
-                    let castp = builder
-                        .build_pointer_cast(tmp, fptr_ty, "i64p_to_f64p")
-                        .map_err(|e| e.to_string())?;
-                    builder.build_store(castp, fv).map_err(|e| e.to_string())?;
-                    builder
-                        .build_load(i64t, tmp, "ld_f2i")
-                        .map_err(|e| e.to_string())?
-                        .into_int_value()
-                }
-                _ => return Err("unsupported value for i64 conversion".to_string()),
-            })
-        }
-        fn i64_to_ptr<'ctx>(
-            ctx: &'ctx Context,
-            builder: &inkwell::builder::Builder<'ctx>,
-            iv: IntValue<'ctx>,
-        ) -> Result<PointerValue<'ctx>, String> {
-            let pty = ctx.i8_type().ptr_type(AddressSpace::from(0));
-            builder
-                .build_int_to_ptr(iv, pty, "i64_to_ptr")
-                .map_err(|e| e.to_string())
-        }
-        fn classify_tag<'ctx>(v: BasicValueEnum<'ctx>) -> i64 {
-            match v {
-                BasicValueEnum::FloatValue(_) => 5,   // float
-                BasicValueEnum::PointerValue(_) => 8, // handle/ptr
-                BasicValueEnum::IntValue(_) => 3,     // integer/bool
-                _ => 3,
-            }
-        }
-        fn to_bool<'ctx>(
-            ctx: &'ctx Context,
-            b: BasicValueEnum<'ctx>,
-            builder: &inkwell::builder::Builder<'ctx>,
-        ) -> Result<IntValue<'ctx>, String> {
-            if let Some(bb) = as_int(b) {
-                // If not i1, compare != 0
-                if bb.get_type().get_bit_width() == 1 {
-                    Ok(bb)
-                } else {
-                    Ok(builder
-                        .build_int_compare(
-                            inkwell::IntPredicate::NE,
-                            bb,
-                            bb.get_type().const_zero(),
-                            "tobool",
-                        )
-                        .map_err(|e| e.to_string())?)
-                }
-            } else if let Some(fv) = as_float(b) {
-                let zero = fv.get_type().const_float(0.0);
-                Ok(builder
-                    .build_float_compare(inkwell::FloatPredicate::ONE, fv, zero, "toboolf")
-                    .map_err(|e| e.to_string())?)
-            } else if let BasicValueEnum::PointerValue(pv) = b {
-                let i64t = ctx.i64_type();
-                let p2i = builder
-                    .build_ptr_to_int(pv, i64t, "p2i")
-                    .map_err(|e| e.to_string())?;
-                Ok(builder
-                    .build_int_compare(inkwell::IntPredicate::NE, p2i, i64t.const_zero(), "toboolp")
-                    .map_err(|e| e.to_string())?)
-            } else {
-                Err("Unsupported value for boolean conversion".to_string())
-            }
-        }
-        fn cmp_eq_ne_any<'ctx>(
-            ctx: &'ctx Context,
-            builder: &inkwell::builder::Builder<'ctx>,
-            op: &crate::mir::CompareOp,
-            lv: BasicValueEnum<'ctx>,
-            rv: BasicValueEnum<'ctx>,
-        ) -> Result<BasicValueEnum<'ctx>, String> {
-            use crate::mir::CompareOp as C;
-            match (lv, rv) {
-                (BasicValueEnum::IntValue(li), BasicValueEnum::IntValue(ri)) => {
-                    let pred = if matches!(op, C::Eq) {
-                        inkwell::IntPredicate::EQ
-                    } else {
-                        inkwell::IntPredicate::NE
-                    };
-                    Ok(builder
-                        .build_int_compare(pred, li, ri, "icmp")
-                        .map_err(|e| e.to_string())?
-                        .into())
-                }
-                (BasicValueEnum::FloatValue(lf), BasicValueEnum::FloatValue(rf)) => {
-                    let pred = if matches!(op, C::Eq) {
-                        inkwell::FloatPredicate::OEQ
-                    } else {
-                        inkwell::FloatPredicate::ONE
-                    };
-                    Ok(builder
-                        .build_float_compare(pred, lf, rf, "fcmp")
-                        .map_err(|e| e.to_string())?
-                        .into())
-                }
-                (BasicValueEnum::PointerValue(_), _) | (_, BasicValueEnum::PointerValue(_)) => {
-                    let li = to_i64_any(ctx, builder, lv)?;
-                    let ri = to_i64_any(ctx, builder, rv)?;
-                    let pred = if matches!(op, C::Eq) {
-                        inkwell::IntPredicate::EQ
-                    } else {
-                        inkwell::IntPredicate::NE
-                    };
-                    Ok(builder
-                        .build_int_compare(pred, li, ri, "pcmp_any")
-                        .map_err(|e| e.to_string())?
-                        .into())
-                }
-                _ => Err("compare type mismatch".to_string()),
-            }
-        }
-
-        // Pre-create allocas for locals on demand (entry-only builder)
-        let mut allocas: HashMap<ValueId, PointerValue> = HashMap::new();
-        let mut entry_builder = codegen.context.create_builder();
-        entry_builder.position_at_end(entry_bb);
-
-        // Helper: map MirType to LLVM basic type (value type)
-        fn map_mirtype_to_basic<'ctx>(
-            ctx: &'ctx Context,
-            ty: &crate::mir::MirType,
-        ) -> BasicTypeEnum<'ctx> {
-            match ty {
-                crate::mir::MirType::Integer => ctx.i64_type().into(),
-                crate::mir::MirType::Float => ctx.f64_type().into(),
-                crate::mir::MirType::Bool => ctx.bool_type().into(),
-                crate::mir::MirType::String => ctx.i8_type().ptr_type(AddressSpace::from(0)).into(),
-                crate::mir::MirType::Box(_)
-                | crate::mir::MirType::Array(_)
-                | crate::mir::MirType::Future(_)
-                | crate::mir::MirType::Unknown => {
-                    ctx.i8_type().ptr_type(AddressSpace::from(0)).into()
-                }
-                crate::mir::MirType::Void => ctx.i64_type().into(), // avoid void as a value type; default to i64
-            }
-        }
-
-        // Helper: create (or get) an alloca for a given pointer-typed SSA value id
-        let mut alloca_elem_types: HashMap<ValueId, BasicTypeEnum> = HashMap::new();
-
-        // Pre-create PHI nodes for all blocks (so we can add incoming from predecessors)
-        let mut phis_by_block: HashMap<
-            crate::mir::BasicBlockId,
-            Vec<(ValueId, PhiValue, Vec<(crate::mir::BasicBlockId, ValueId)>)>,
-        > = HashMap::new();
-        for bid in func.block_ids() {
-            let bb = *bb_map.get(&bid).ok_or("missing bb in map")?;
-            // Position at start of the block (no instructions emitted yet)
-            codegen.builder.position_at_end(bb);
-            let block = func.blocks.get(&bid).unwrap();
-            for inst in block
-                .instructions
-                .iter()
-                .take_while(|i| matches!(i, MirInstruction::Phi { .. }))
-            {
-                if let MirInstruction::Phi { dst, inputs } = inst {
-                    // Decide PHI type: prefer annotated value type; fallback to first input's annotated type; finally i64
-                    let mut phi_ty: Option<BasicTypeEnum> = None;
-                    if let Some(mt) = func.metadata.value_types.get(dst) {
-                        phi_ty = Some(map_mirtype_to_basic(codegen.context, mt));
-                    } else if let Some((_, iv)) = inputs.first() {
-                        if let Some(mt) = func.metadata.value_types.get(iv) {
-                            phi_ty = Some(map_mirtype_to_basic(codegen.context, mt));
-                        }
-                    }
-                    let phi_ty = phi_ty.unwrap_or_else(|| codegen.context.i64_type().into());
-                    let phi = codegen
-                        .builder
-                        .build_phi(phi_ty, &format!("phi_{}", dst.as_u32()))
-                        .map_err(|e| e.to_string())?;
-                    vmap.insert(*dst, phi.as_basic_value());
-                    phis_by_block
-                        .entry(bid)
-                        .or_default()
-                        .push((*dst, phi, inputs.clone()));
-                }
-            }
-        }
-
-        // Lower in block order
-        for bid in func.block_ids() {
-            let bb = *bb_map.get(&bid).unwrap();
-            if codegen
-                .builder
-                .get_insert_block()
-                .map(|b| b != bb)
-                .unwrap_or(true)
-            {
-                codegen.builder.position_at_end(bb);
-            }
-            let block = func.blocks.get(&bid).unwrap();
-            for inst in &block.instructions {
+pub(crate) fn lower_instruction<'ctx>(
+    inst: &MirInstruction,
+    codegen: &CodegenContext<'ctx>,
+    func: &MirFunction,
+    bid: BasicBlockId,
+    vmap: &mut HashMap<ValueId, BasicValueEnum<'ctx>>,
+    allocas: &mut HashMap<ValueId, PointerValue<'ctx>>,
+    entry_builder: &Builder<'ctx>,
+    alloca_elem_types: &mut HashMap<ValueId, BasicTypeEnum<'ctx>>,
+    phis_by_block: &HashMap<BasicBlockId, Vec<(ValueId, PhiValue<'ctx>, Vec<(BasicBlockId, ValueId)>)>>,
+    bb_map: &HashMap<BasicBlockId, BasicBlock<'ctx>>,
+    box_type_ids: &HashMap<String, i64>,
+) -> Result<(), String> {
                 match inst {
                     MirInstruction::NewBox {
                         dst,
@@ -2434,97 +2135,5 @@ impl LLVMCompiler {
                     }
                     _ => {}
                 }
-            }
-        }
-
-        // Verify and emit
-        if !llvm_func.verify(true) {
-            return Err("Function verification failed".to_string());
-        }
-        // Try writing via file API first; if it succeeds but file is missing due to env/FS quirks,
-        // also write via memory buffer as a fallback to ensure presence.
-        let verbose = std::env::var("NYASH_CLI_VERBOSE").ok().as_deref() == Some("1");
-        if verbose {
-            eprintln!("[LLVM] emitting object to {} (begin)", output_path);
-        }
-        match codegen.target_machine.write_to_file(
-            &codegen.module,
-            inkwell::targets::FileType::Object,
-            std::path::Path::new(output_path),
-        ) {
-            Ok(_) => {
-                // Verify; if missing, fallback to memory buffer write
-                if std::fs::metadata(output_path).is_err() {
-                    let buf = codegen
-                        .target_machine
-                        .write_to_memory_buffer(&codegen.module, inkwell::targets::FileType::Object)
-                        .map_err(|e| format!("Failed to get object buffer: {}", e))?;
-                    std::fs::write(output_path, buf.as_slice()).map_err(|e| {
-                        format!("Failed to write object to '{}': {}", output_path, e)
-                    })?;
-                    if verbose {
-                        eprintln!(
-                            "[LLVM] wrote object via memory buffer fallback: {} ({} bytes)",
-                            output_path,
-                            buf.get_size()
-                        );
-                    }
-                } else if verbose {
-                    if let Ok(meta) = std::fs::metadata(output_path) {
-                        eprintln!(
-                            "[LLVM] wrote object via file API: {} ({} bytes)",
-                            output_path,
-                            meta.len()
-                        );
-                    }
-                }
-                if verbose {
-                    eprintln!("[LLVM] emit complete (Ok branch) for {}", output_path);
-                }
-                Ok(())
-            }
-            Err(e) => {
-                // Fallback: memory buffer
-                let buf = codegen
-                    .target_machine
-                    .write_to_memory_buffer(&codegen.module, inkwell::targets::FileType::Object)
-                    .map_err(|ee| {
-                        format!(
-                            "Failed to write object ({}); and memory buffer failed: {}",
-                            e, ee
-                        )
-                    })?;
-                std::fs::write(output_path, buf.as_slice()).map_err(|ee| {
-                    format!(
-                        "Failed to write object to '{}': {} (original error: {})",
-                        output_path, ee, e
-                    )
-                })?;
-                if verbose {
-                    eprintln!(
-                        "[LLVM] wrote object via error fallback: {} ({} bytes)",
-                        output_path,
-                        buf.get_size()
-                    );
-                }
-                if verbose {
-                    eprintln!(
-                        "[LLVM] emit complete (Err branch handled) for {}",
-                        output_path
-                    );
-                }
-                Ok(())
-            }
-        }
-    }
-}
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_compiler_creation() {
-        let compiler = LLVMCompiler::new();
-        assert!(compiler.is_ok());
-    }
+    Ok(())
 }
